requirements.txt
# Living Library - Dependencies

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9 
supabase==2.3.4
PyMuPDF==1.24.14
sentence-transformers==2.3.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
httpx
pgvector==0.2.4
rich==14.0.0
